+++ NEWS.md
@@ -2,9 +2,11 @@
 
 ##### Changes
 * add support for rds files (single object data files from R) [#22], [#33]
+* add support for Date and POSIXct, though still lacking timezone handling [#34]
 
 [#22]: https://github.com/JuliaStats/RData.jl/issues/22
 [#33]: https://github.com/JuliaStats/RData.jl/issues/33
+[#34]: https://github.com/JuliaStats/RData.jl/issues/34
 
 ## RData v0.2.0 Release Notes